requirements.txt
Django==5.1.2
python-decouple==3.8
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0
django-sesame==3.2.2
celery==5.3.4
redis==5.0.1
fire==0.5.0
daphne==4.0.0
logfire
stripe==8.0.0 